fixes the u8x2 visibility bug
